+++ app/services/lab/lims/api/rest_api.rb
@@ -22,14 +22,13 @@ class Lab::Lims::Api::RestApi
       end
     end
 
-    update_order_results(order_dto)
-
-    data = JSON.parse(response.body)['data']
+    data = JSON.parse(response.body)
+    update_order_results(order_dto) unless data['message'].casecmp?('Order already available')
 
     ActiveSupport::HashWithIndifferentAccess.new(
-      id: data['tracking_number'],
+      id: order_dto.fetch(:_id, order_dto[:tracking_number]),
       rev: 0,
-      tracking_number: data['tracking_number']
+      tracking_number: order_dto[:tracking_number]
     )
   end
 
@@ -44,21 +43,22 @@ class Lab::Lims::Api::RestApi
   end
 
   def consume_orders(*_args, patient_id: nil, **_kwargs)
-    Rails.logger.info('Looking for orders without results...')
-    orders = Lab::OrdersSearchService.find_orders_without_results(patient_id: patient_id)
+    orders_pending_updates(patient_id).each do |order|
+      order_dto = Lab::Lims::OrderSerializer.serialize_order(order)
 
-    orders.each do |order|
-      response = in_authenticated_session do |headers|
-        Rails.logger.info("Fetching results for order ##{order.accession_number}")
-        RestClient.get(expand_uri("query_results_by_tracking_number/#{order.accession_number}"), headers)
+      if order_dto['priority'].nil? || order_dto['sample_type'].casecmp?('not_specified')
+        patch_order_dto_with_lims_order!(order_dto, find_lims_order(order.accession_number))
       end
 
-      Rails.logger.info("Result for order ##{order.accession_number} found... Parsing...")
-      results = JSON.parse(response).fetch('data').fetch('results')
-      order_dto = Lab::Lims::OrderSerializer.serialize_order(order)
-      yield lims_results_to_order_dto(results, order_dto), OpenStruct.new(last_seq: 0)
-    rescue InvalidParameters => e # LIMS responds with a 401 when a result is not found :(
-      Rails.logger.error("Failed to fetch results for ##{order.accession_number}: #{e.message}")
+      if order_dto['test_results'].empty?
+        begin
+          patch_order_dto_with_lims_results!(order_dto, find_lims_results(order.accession_number))
+        rescue InvalidParameters => e # LIMS responds with a 401 when a result is not found :(
+          Rails.logger.error("Failed to fetch results for ##{order.accession_number}: #{e.message}")
+        end
+      end
+
+      yield order_dto, OpenStruct.new(last_seq: 0)
     end
   end
 
@@ -105,8 +105,8 @@ class Lab::Lims::Api::RestApi
     self.authentication_token = nil
     retry if (retries -= 1).positive?
   rescue RestClient::ExceptionWithResponse => e
-    Rails.logger.error("LIMS Error: #{e.response.code} - #{e.response.body}")
-    raise e unless e.response.code == 401
+    Rails.logger.error("LIMS Error: #{e.response&.code} - #{e.response&.body}")
+    raise e unless e.response&.code == 401
 
     self.authentication_token = nil
     retry if (retries -= 1).positive?
@@ -117,7 +117,8 @@ class Lab::Lims::Api::RestApi
     password = config.fetch(:password)
 
     Rails.logger.debug("Authenticating with LIMS as: #{username}")
-    response = RestClient.get(expand_uri("re_authenticate/#{username}/#{password}"), headers: { 'Content-type' => 'application/json' })
+    response = RestClient.get(expand_uri("re_authenticate/#{username}/#{password}"),
+                              headers: { 'Content-type' => 'application/json' })
     response_body = JSON.parse(response.body)
 
     if response_body['status'] == 401
@@ -153,9 +154,7 @@ class Lab::Lims::Api::RestApi
 
     Rails.logger.error("Lims Api Error: #{response.body}")
 
-    if body['status'] != 401
-      raise LimsApiError, "#{body['status']} - #{body['message']}"
-    end
+    raise LimsApiError, "#{body['status']} - #{body['message']}" if body['status'] != 401
 
     if body['message'].match?(/token expired/i)
       raise AuthenticationTokenExpired, "Authentication token expired: #{body['message']}"
@@ -255,11 +254,31 @@ class Lab::Lims::Api::RestApi
     "#{orderer[:first_name]} #{orderer[:last_name]}"
   end
 
+  def find_lims_order(tracking_number)
+    response = in_authenticated_session do |headers|
+      Rails.logger.info("Fetching order ##{tracking_number}")
+      RestClient.get(expand_uri("query_order_by_tracking_number/#{tracking_number}"), headers)
+    end
+
+    Rails.logger.info("Order ##{tracking_number} found... Parsing...")
+    JSON.parse(response).fetch('data')
+  end
+
+  def find_lims_results(tracking_number)
+    response = in_authenticated_session do |headers|
+      Rails.logger.info("Fetching results for order ##{tracking_number}")
+      RestClient.get(expand_uri("query_results_by_tracking_number/#{tracking_number}"), headers)
+    end
+
+    Rails.logger.info("Result for order ##{tracking_number} found... Parsing...")
+    JSON.parse(response).fetch('data').fetch('results')
+  end
+
   ##
   # Make a copy of the order_dto with the results from LIMS parsed
   # and appended to it.
-  def lims_results_to_order_dto(results, order_dto)
-    order_dto.merge(
+  def patch_order_dto_with_lims_results!(order_dto, results)
+    order_dto.merge!(
       '_id' => order_dto[:tracking_number],
       '_rev' => 0,
       'test_results' => results.each_with_object({}) do |result, formatted_results|
@@ -277,10 +296,16 @@ class Lab::Lims::Api::RestApi
     )
   end
 
+  def patch_order_dto_with_lims_order!(order_dto, lims_order)
+    order_dto.merge!(
+      'sample_type' => lims_order['other']['sample_type'],
+      'sample_status' => lims_order['other']['specimen_status'],
+      'priority' => lims_order['other']['priority']
+    )
+  end
+
   def update_order_results(order_dto)
-    if order_dto['test_results'].nil? || order_dto['test_results'].empty?
-      return nil
-    end
+    return nil if order_dto['test_results'].nil? || order_dto['test_results'].empty?
 
     order_dto['test_results'].each do |test_name, results|
       Rails.logger.info("Pushing result for order ##{order_dto['tracking_number']}")
@@ -340,5 +365,41 @@ class Lab::Lims::Api::RestApi
       },
       test_status: 'voided'
     }
+  end
+
+  def orders_pending_updates(patient_id = nil)
+    Rails.logger.info('Looking for orders that need to be updated...')
+    orders = {}
+
+    orders_without_specimen(patient_id).each { |order| orders[order.order_id] = order }
+    orders_without_results(patient_id).each { |order| orders[order.order_id] = order }
+    orders_without_reason(patient_id).each { |order| orders[order.order_id] = order }
+
+    orders.values
+  end
+
+  def orders_without_specimen(patient_id = nil)
+    Rails.logger.debug('Looking for orders without a specimen')
+    unknown_specimen = ConceptName.where(name: Lab::Metadata::UNKNOWN_SPECIMEN)
+                                  .select(:concept_id)
+    orders = Lab::LabOrder.where(concept_id: unknown_specimen)
+    orders = orders.where(patient_id: patient_id) if patient_id
+
+    orders
+  end
+
+  def orders_without_results(patient_id = nil)
+    Rails.logger.debug('Looking for orders without a result')
+    Lab::OrdersSearchService.find_orders_without_results(patient_id: patient_id)
+  end
+
+  def orders_without_reason(patient_id = nil)
+    Rails.logger.debug('Looking for orders without a reason for test')
+    orders = Lab::LabOrder.joins(:reason_for_test)
+                          .merge(Observation.where(value_coded: nil, value_text: nil))
+                          .limit(1000)
+    orders = orders.where(patient_id: patient_id) if patient_id
+
+    orders
   end
 end